Fixed issues with strncmp and parameter validation

+++ Userland/main_app/programs/shell.c
@@ -53,6 +53,30 @@ static void createWindow(){
 }
 
 
+void consumer(int argc, char **argv){
+
+	static int x=0;
+
+	block(getPid());
+
+	for(; x<x+100; x++)
+			draw(x, 100, 0xFF0000);
+
+	// while(1){
+	// 	waitSem(0);
+	// 	for(; x<x+100; x++)
+	// 		draw(x, 100, 0xFF0000);
+	// }
+}
+
+void producer(int argc, char **argv){
+	while(1){
+		for(int i=0; i<9999; i++);
+		postSem(0);
+	}
+}
+
+
 void shell(){
 
 	createWindow();
@@ -62,6 +86,13 @@ void shell(){
 	int buffPos = 0;
 
 	w.activeCursor = bodyCursor;
+
+	if(createSem(0, 0) == -1){
+		printf("Failed to create semaphore\n", 0);
+	}
+
+	// startProcess(consumer, 0, NULL, "consumer");
+	// startProcess(producer, 0, NULL, "producer");
 
 
 	char c;
@@ -111,7 +142,8 @@ void parseCommand(char *cmdBuff){
 	char *tokens[MAX_TOKENS] = {0};
 	tokens[0] = cmdBuff;
 
-	for (int i=0, j=1; cmdBuff[i]; i++){
+	int j=1;
+	for (int i=0; cmdBuff[i]; i++){
 		if(isSpace(cmdBuff[i])){
 			cmdBuff[i] = 0;
 			tokens[j++] = cmdBuff+i+1;
@@ -120,67 +152,67 @@ void parseCommand(char *cmdBuff){
 
 	// Miscellaneous
 
-	if(strncmp(tokens[0], "help", 5))
+	if(strncmp(tokens[0], "help", 5) == 0)
 		printHelp();
 
-	else if(strncmp(tokens[0], "divzero", 8))
+	else if(strncmp(tokens[0], "divzero", 8) == 0)
 		divZeroException();
 
-	else if(strncmp(tokens[0], "invopcode", 10))
+	else if(strncmp(tokens[0], "invopcode", 10) == 0)
 		invOpcodeException();
 
-	else if(strncmp(tokens[0], "time", 5))
+	else if(strncmp(tokens[0], "time", 5) == 0)
 		printTime();
 
-	else if(strncmp(tokens[0], "display", 8))
+	else if(strncmp(tokens[0], "display", 8) == 0 && j==2)
 		displayImage(tokens[1], 20, 200);
 	
-	else if(strncmp(tokens[0], "clear", 6))
+	else if(strncmp(tokens[0], "clear", 6) == 0)
 		clearWindow();
 
 
 	// CPU management
 
-	else if(strncmp(tokens[0], "cputemp", 8))
+	else if(strncmp(tokens[0], "cputemp", 8) == 0)
 		printCPUTemp();
 
-	else if(strncmp(tokens[0], "cpuinfo", 8))
+	else if(strncmp(tokens[0], "cpuinfo", 8) == 0)
 		printCPUInfo();
 
-	else if(strncmp(tokens[0], "inforeg", 8))
+	else if(strncmp(tokens[0], "inforeg", 8) == 0)
 		printInfoReg();
 
-	else if(strncmp(tokens[0], "storedreg", 10))
+	else if(strncmp(tokens[0], "storedreg", 10) == 0)
 		printStoredReg();
 	
 
 	// Memory management
 
-	else if(strncmp(tokens[0], "memdump", 8))
+	else if(strncmp(tokens[0], "memdump", 8) == 0 && j==2)
 		printMemDump(tokens[1]);
 	
-	else if(strncmp(tokens[0], "heapstat", 9))
+	else if(strncmp(tokens[0], "heapstat", 9) == 0)
 		printMemStatus();
 
 	
 	// Process management
 	
-	else if(strncmp(tokens[0], "ps", 3))
+	else if(strncmp(tokens[0], "ps", 3) == 0)
 		printProcData();
 
-	else if(strncmp(tokens[0], "kill", 5))
+	else if(strncmp(tokens[0], "kill", 5) == 0 && j==2)
 		killProcess(tokens[1]);
 
-	else if(strncmp(tokens[0], "block", 6))
+	else if(strncmp(tokens[0], "block", 6) == 0 && j==2)
 		blockProcess(tokens[1]);
 	
 
 	// New processes
 
-	else if(strncmp(tokens[0], "line", 5))	
+	else if(strncmp(tokens[0], "line", 5) == 0)	
 		startProcess(drawLine, 0, NULL, "line");
 
-	else if(strncmp(tokens[0], "loop", 5))
+	else if(strncmp(tokens[0], "loop", 5) == 0)
 		startProcess(loop, 0, NULL, "loop");
 
 
@@ -199,7 +231,7 @@ void parseCommand(char *cmdBuff){
 void clearWindow(){
 	for (int x = 0; x < w.xf; x++)
 	{
-		for (int y = bodyY; y < w.yf; y++)
+		for (int y = 0; y < w.yf; y++)
 		{
 			//draw(x, y, 0);
 			drawPoint(x, y, 1, 0);
@@ -208,24 +240,6 @@ void clearWindow(){
 
 	w.cursors[bodyCursor].x = 0;
 	w.cursors[bodyCursor].y = bodyY;
-}
-
-
-void consumer(int argc, char **argv){
-	static int x=0;
-
-	while(1){
-		waitSem(0);
-		for(; x<x+100; x++)
-			draw(x, 100, 0xFF0000);
-	}
-}
-
-void producer(int argc, char **argv){
-	while(1){
-		for(int i=0; i<9999; i++);
-		postSem(0);
-	}
 }
